CI that gates on something, and the server-side request forgery turning it on found
The discipline job is called "the rules that hold the codebase to its own
claims" and ran no linter, no type checker, no coverage floor and nothing that
looked at dependencies. There is now a hygiene job that runs first because it is
the cheapest signal.

**The linter found a real defect on its first pass.** `token_endpoint` and
`jwks_uri` are read out of an identity provider's own discovery document and
then fetched, and nothing checked the scheme — `urllib.request.urlopen` handles
`file:` as readily as `https:`, so an issuer returning
`"jwks_uri": "file:///etc/shadow"` had the platform open it. `core/outbound.py`
decides in one place and fails closed: https, or http to loopback where every
resolved address is loopback, and an unresolvable host is not loopback because
failing open there would make every typo a permitted plain-http fetch.

The rule set is chosen rather than inherited, and pyproject.toml says why each
rule is in or out. The default reports three thousand findings, nearly all of
them that this codebase writes `Dict[str, Any]`; a linter reporting three
thousand findings is one nobody acts on. ISC004 fires 195 times on correct
multi-line f-strings and is out for the same reason. What is in is the set that
finds defects — and it found two more: a closure over a loop variable in a
property test, so a failure in any round but the last would have been evaluated
against the wrong data, and a dead re-read in the lifecycle service.

Types gate on the 186 modules that check cleanly, with 62 in a backlog file.
`mypy || true` is a step that always passes — the defect this codebase is named
for — and `--strict` across 248 modules produces a blanket ignore, which is the
same step wearing a hat. The dominant cause is one honest lie:
`Repository.one()` returns Optional and a caller that just wrote the row
annotates otherwise. Repairing that is a hundred call sites and its own release.

Coverage is a floor at 90, chosen after measuring rather than before: the suite
sits at 93, and the 80 I first wrote would have caught nothing.

`openapi.lock.json` locks the API's shape — paths, methods, parameter and body
field names, requiredness — because no test here notices a renamed route or a
body that gains a required key. Deliberately not the whole document: locking
descriptions would make a reworded docstring an interface change.

pip-audit runs against requirements.txt rather than the environment. Its first
run reported seven advisories, every one against pip itself, and a security gate
whose first output is about the toolchain is one people learn to skim.

**The secret scanner shipped reporting success while doing nothing**, which is
the defect it was written to help catch, in the tool written to catch it:
`any(part in SKIP_DIRS for part in SKIP_DIRS)` asks whether each member of a set
is in that set, so every file was skipped and it reported "no secrets found in
551 tracked files" having read none of them. Found by planting a private key and
watching it pass. There is a test that plants one now.

Also: my own `RUF100 --fix` stripped six deliberate `noqa: F401` comments before
`F401` ran and removed the imports under them, breaking a re-export eighteen
tests depended on — and the F821 that came with it was caught by the linter I
had just installed, one command later.

2,896 passing, 93% covered, all seven gates green.

+++ core/validation/statistics.py
@@ -113,7 +113,7 @@ def psi(expected: Sequence[float], actual: Sequence[float], bins: int = 10) -> O
     if len(expected) < bins or not actual:
         return None
     ordered = sorted(expected)
-    edges = [ordered[int(round(q * (len(ordered) - 1) / bins))] for q in range(1, bins)]
+    edges = [ordered[round(q * (len(ordered) - 1) / bins)] for q in range(1, bins)]
     floor = 1.0 / (bins * 10.0)
 
     def shares(sample: Sequence[float]) -> List[float]:
